Move testing app function to app helpers
