+++ app/app.go
@@ -50,7 +50,6 @@ import (
 	txmodule "github.com/cosmos/cosmos-sdk/x/auth/tx/config"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	"github.com/cosmos/cosmos-sdk/x/crisis"
-	govkeeper "github.com/cosmos/cosmos-sdk/x/gov/keeper"
 	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
 
 	"github.com/bandprotocol/chain/v3/app/keepers"
@@ -507,21 +506,4 @@ func (app *BandApp) AutoCliOpts() autocli.AppOptions {
 		ValidatorAddressCodec: authcodec.NewBech32Codec(sdk.GetConfig().GetBech32ValidatorAddrPrefix()),
 		ConsensusAddressCodec: authcodec.NewBech32Codec(sdk.GetConfig().GetBech32ConsensusAddrPrefix()),
 	}
-}
-
-// TestingApp functions
-
-// GetBaseApp implements the TestingApp interface.
-func (app *BandApp) GetBaseApp() *baseapp.BaseApp {
-	return app.BaseApp
-}
-
-// GetTxConfig implements the TestingApp interface.
-func (app *BandApp) GetTxConfig() client.TxConfig {
-	return app.txConfig
-}
-
-// GetTestGovKeeper implements the TestingApp interface.
-func (app *BandApp) GetTestGovKeeper() *govkeeper.Keeper {
-	return app.AppKeepers.GovKeeper
 }
